% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy_est_functions.R
\name{area_est_ploidy_single_sample}
\alias{area_est_ploidy_single_sample}
\title{Estimate ploidy using area method}
\usage{
area_est_ploidy_single_sample(data_sample, ploidys, area)
}
\arguments{
\item{data_sample}{data.frame with BAF values and genomic position information}

\item{ploidys}{range of ploidies to by tested}

\item{area}{area around the expected peak to be considered}
}
\description{
Estimate ploidy using area method
}
