% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy_est.R
\name{area_estimate_ploidy}
\alias{area_estimate_ploidy}
\title{Estimate ploidy using area method}
\usage{
area_estimate_ploidy(
  qploidy_standardization = NULL,
  samples = "all",
  level = "chromosome",
  ploidies = NULL,
  area = 0.75,
  centromeres = NULL
)
}
\arguments{
\item{qploidy_standardization}{object of class qploidy_standardization}

\item{samples}{if "all" all samples contained in the qploidy_standardization object will be evaluate. If vector with sample names is provided, only those will be evaluated.}

\item{level}{character identifying the level of the analysis. If `chromosome` the number of copies will be estimated by chromosome, if `sample` it will be by sample; if `chromosome-arm` it will be peformed by chromosome arm (only if `centromeres` argument is defined.)}

\item{ploidies}{range of ploidies to by tested}

\item{area}{area around the expected peak to be considered}

\item{centromeres}{vector with centromeres genomic position in bp. The vector should be named with the chromosomes ID. The information will only be used if `chromosome-arm` level is defined.}
}
\description{
Estimate ploidy using area method
}
