% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy_est_functions.R
\name{area_estimate_ploidy_by_chr}
\alias{area_estimate_ploidy_by_chr}
\title{Estimate ploidy using area method}
\usage{
area_estimate_ploidy_by_chr(data_sample, ploidys, area)
}
\description{
Estimate ploidy using area method
}
