% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{clean_summary}
\alias{clean_summary}
\title{Clean Axiom Summary File}
\usage{
clean_summary(summary_df)
}
\arguments{
\item{summary_df}{A data frame containing A and B probe intensities.}
}
\value{
A list with cleaned A and B probes.
}
\description{
This function removes consecutive A allele probes from an Axiom summary 
file.
}
\examples{
NULL
}
