% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_counts_functions.R
\name{count_breaks_df}
\alias{count_breaks_df}
\title{Count number of recombination breakpoints in parents gametes}
\usage{
count_breaks_df(homoprob, aneuploids = NULL, inds = NULL, ped.file = NULL)
}
\arguments{
\item{homoprob}{object of class mappoly.homoprob}

\item{aneuploids}{data.frame with each chromosome ploidy}

\item{inds}{vector with individuals to be evaluated}

\item{ped.file}{path to csv file containing three columns: 1) Sample ID, 2) Parent1 ID, 3) Parent2 ID}
}
\description{
Count number of recombination breakpoints in parents gametes
}
