% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_counts_functions.R
\name{count_breaks_mappoly}
\alias{count_breaks_mappoly}
\title{Count number of recombination breakpoints in parents gametes}
\usage{
count_breaks_mappoly(
  homoprob,
  aneuploids = NULL,
  n.graphics = NULL,
  ncol = NULL,
  by_LG = TRUE
)
}
\arguments{
\item{n.graphics}{define number of graphics}

\item{ncol}{define number of columns}

\item{object}{of class mappoly.homoprob}
}
\description{
Count number of recombination breakpoints in parents gametes
}
