% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_counts.R
\name{count_breaks_plot}
\alias{count_breaks_plot}
\title{Plot break counts}
\usage{
count_breaks_plot(counts, n.graphics = NULL, ncol = NULL, by_LG = TRUE)
}
\arguments{
\item{counts}{data.frame generated by count_breaks_df}

\item{n.graphics}{define number of graphics}

\item{ncol}{define number of columns}

\item{by_LG}{FALSE if colors by ploidy, TRUE if colors by group}
}
\description{
Plot break counts
}
