% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QC_plots.R
\name{depth_pca_plot}
\alias{depth_pca_plot}
\title{PCA Plot for Qploidy Standardized Data}
\usage{
depth_pca_plot(
  input,
  geno.pos = NULL,
  passport_file,
  group_column,
  sampleID_column,
  col2use = c("R", "z")
)
}
\arguments{
\item{input}{Accepts one of:
1. A data.frame with columns: MarkerName, SampleName, Chr, Position, R. Column with z is optional but required if col2use = "z".
2. An object of class 'qploidy_standardization' (from Qploidy standardization functions)
3. A character string: path to a .tsv[.gz] file with standardized Qploidy data
The function will validate and process the input according to its type.}

\item{geno.pos}{Optional. A data.frame with columns: MarkerName, Chromosome, Position. Required if input data.frame does not contain Chr and Position columns.}

\item{passport_file}{Either a data.frame with sample metadata (first column must contain sample IDs matching the Qploidy object), or a character path to a CSV or TXT file containing sample metadata.}

\item{group_column}{Character. The name of the column in the passport file
used to color/group the PCA plots.}

\item{sampleID_column}{character or numeric indicating the column name or index where the sample name is located. Sample name must match `input`}

\item{col2use}{Character. Either "R" or "z". Which Qploidy column to use for
the PCA. Defaults to c("R","z").}
}
\value{
Prints three ggplot2 PCA plots: PC1 vs PC2, PC1 vs PC3, and PC2 vs PC3.
  Returns a list of these plots invisibly.
}
\description{
Performs Principal Component Analysis (PCA) on a Qploidy standardized object
using either the "R" or "z" column, and colors the samples based on a grouping
variable from a passport file (CSV or TXT).
}
\examples{
depth_pca_plot(
  qploidy_object = "standardization.tsv.gz",
  passport_file = "pca_ids.csv",
  group_column = "Plate",
  col2use = "R"
)
}
\author{
Josue Chinchilla-Vargas
}
