% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dosages.R
\name{export_VCF}
\alias{export_VCF}
\title{Export hmm_dosage_calls to VCF}
\usage{
export_VCF(hmm_dosage_calls, file)
}
\arguments{
\item{hmm_dosage_calls}{A data.frame with columns: MarkerName, SampleName, Chr, Position, X, Y, CN_call, post_max_CN, dosage, post_max_dosage, BAF, z}

\item{file}{Path to output VCF file}
}
\description{
Converts a hmm_dosage_calls object (data.frame) to a VCF file with GT, CN, AD, BAF, Z, post_max_CN, and post_max_dosage in FORMAT.
}
