% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{find_header_line}
\alias{find_header_line}
\title{Find the Header Line in a File}
\usage{
find_header_line(summary_file, word = "probeset_id", max_lines = 6000)
}
\arguments{
\item{summary_file}{The path to the file to be scanned.}

\item{word}{The keyword to search for in the first column. Default is
"probeset_id".}

\item{max_lines}{The maximum number of lines to scan. Default is 6000.}
}
\value{
The line number where the keyword is found.
}
\description{
This function scans a file to locate the first line containing a specific
keyword, such as 'probeset_id'. It is useful for identifying the starting
point of data in files with headers or metadata.
}
