% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy_est_functions.R
\name{get_aneuploids}
\alias{get_aneuploids}
\title{indexes for aneuploids}
\usage{
get_aneuploids(ploidy_df)
}
\arguments{
\item{ploidy_df}{ploidy table (chromosome in columns and individuals in rows)}
}
\description{
indexes for aneuploids
}
