% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{get_baf_par}
\alias{get_baf_par}
\title{To create baf in parallel}
\usage{
get_baf_par(par_all_item, ploidy = 2)
}
\arguments{
\item{par_all_item}{list containing R and theta matrices, and clusters models}

\item{ploidy}{integer defining ploidy}
}
\description{
To create baf in parallel
}
