% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{get_logR_par}
\alias{get_logR_par}
\title{To create logR in parallel}
\usage{
get_logR_par(par_all_item, ploidy = 2)
}
\arguments{
\item{par_all_item}{list containing R and theta matrices, and clusters models}

\item{ploidy}{integer defining ploidy}
}
\description{
To create logR in parallel
}
