% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{get_probs_polyorigin}
\alias{get_probs_polyorigin}
\title{Converts polyOrigin genotype probabilities format to mappoly homoprob format}
\usage{
get_probs_polyorigin(df, f1.codes, ind = NULL, ploidy = 4, n.cores = 1)
}
\arguments{
\item{df}{data.frame resulted from read_polyOrigin function}

\item{f1.codes}{data.frame with polyOrigin stages codification}

\item{ind}{character defining the individual to be evalueated}

\item{ploidy}{integer defining the ploidy}

\item{n.cores}{define how many cores to be used in parallelization}
}
\description{
Converts polyOrigin genotype probabilities format to mappoly homoprob format
}
