% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{get_probs_sing}
\alias{get_probs_sing}
\title{Converts polyOrigin genotype probabilities format to mappoly homoprob format for a single individual}
\usage{
get_probs_sing(melt_df, f1.codes, ploidy)
}
\arguments{
\item{melt_df}{long format of the data.frame resulted from read_polyOrigin function}

\item{f1.codes}{data.frame with polyOrigin stages codification}

\item{ploidy}{integer defining the ploidy}
}
\description{
Converts polyOrigin genotype probabilities format to mappoly homoprob format for a single individual
}
