% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{get_R_theta_illumina}
\alias{get_R_theta_illumina}
\title{Get R and theta values from Illumina intensities file with option to do standard normalization by plate and markers}
\usage{
get_R_theta_illumina(cleaned_illumina, atan = FALSE)
}
\arguments{
\item{cleaned_illumina}{List with X and Y intensities.}

\item{atan}{logical. If TRUE calculates the theta using atan2}
}
\description{
Get R and theta values from Illumina intensities file with option to do standard normalization by plate and markers
}
