% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy_est_functions.R
\name{merge_arms_format}
\alias{merge_arms_format}
\title{Merges chromosome-arm level analysis results into chromosome level format}
\usage{
merge_arms_format(x, filter_diff = NULL)
}
\arguments{
\item{x}{object of class qploidy_area_ploidy_estimation}

\item{filter_diff}{filter by difference on area proportion between first and second place}
}
\description{
Merges chromosome-arm level analysis results into chromosome level format
}
