% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode}
\alias{mode}
\title{Calculate the Statistical Mode}
\usage{
mode(x)
}
\arguments{
\item{x}{A vector of numeric, character, or factor values.}
}
\value{
A single value representing the mode of the input vector.
}
\description{
This function returns the most frequent (modal) value in a vector.
If there are multiple values with the same highest frequency,
it returns the first one encountered.
}
\examples{
mode(c(1, 2, 2, 3, 4))   # Returns 2
mode(c("apple", "banana", "apple"))  # Returns "apple"

}
