% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{plot.qploidy_interpolation}
\alias{plot.qploidy_interpolation}
\title{PLot method for object of class 'qploidy_interpolation'}
\usage{
\method{plot}{qploidy_interpolation}(
  x,
  sample = NULL,
  chr = NULL,
  type = c("all", "het", "ratio", "BAF", "zscore", "BAF_hist", "BAF_hist_overall"),
  area_single = 0.75,
  ploidy = 4,
  dot.size = 1,
  add_estimated_peaks = FALSE,
  add_expected_peaks = FALSE,
  centromeres = NULL,
  add_centromeres = FALSE,
  colors = FALSE,
  window_size = 2e+06
)
}
\arguments{
\item{x}{object of class 'qploidy_interpolation'}

\item{sample}{character indicating sample ID}

\item{type}{character defining the graphic type. "all" plots all graphics,
it is equivalent to "het","ratio","BAF","zscore","BAF_hist", "BAF_hist_overall".
"het" plots the heterozygous locus counts inside genomic positions window defined in 'window_size'.
"ratio" plots raw Y/(X+Y) or alternative count/(alternative counts + reference counts).
"BAF" plots interpolated ratios. "zscore" is the smoothed conditional means curve of standardized sum of intensities/counts.
"BAF_hist" is the histogram of interpolated ratios. "BAF_hist_overall" add the histogram including all markers.}

\item{area_single}{area around the expected peak to be considered}

\item{ploidy}{expected ploidy}

\item{add_estimated_peaks}{add expected peaks lines}

\item{add_expected_peaks}{add estimated peaks}

\item{colors}{add area colors}

\item{window_size}{genomic position window to calculate the number of heterozygous locus}
}
\value{
printed information about Qploidy interpolation process
}
\description{
PLot method for object of class 'qploidy_interpolation'
}
