% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy_est_functions.R
\name{plot_baf}
\alias{plot_baf}
\title{Plot BAF}
\usage{
plot_baf(data_sample, area_single, ploidy, dot.size, add_lines, colors)
}
\arguments{
\item{data_sample}{data.frame with BAF and genomic position information}

\item{area_single}{area around the expected peak to be considered}

\item{ploidy}{expected ploidy}

\item{dot.size}{graphic dot size}

\item{add_lines}{add expected peaks lines}

\item{colors}{add area colors}
}
\description{
Plot BAF
}
