% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_check}
\alias{plot_check}
\title{3D Plot for Raw Data Inspection}
\usage{
plot_check(data_qploidy, R_lim = NULL, n = 3000)
}
\arguments{
\item{data_qploidy}{A data.frame input for Qploidy. Must include columns `MarkerName`, `SampleName`, and `R`.}

\item{R_lim}{Numeric value to limit the sum of intensities/depth axis. Useful for adjusting dimensions in case of outliers. Default is NULL.}

\item{n}{Integer specifying the subset size for plotting. Default is 3000.}
}
\value{
A plotly object representing the 3D surface plot.
}
\description{
This function generates a 3D surface plot to inspect raw data properties, such as allele intensities or sequencing depth. It helps identify variation among individuals for the same marker.
}
