% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{plot_one_marker}
\alias{plot_one_marker}
\title{Interpolation plot}
\usage{
plot_one_marker(scores_temp, ploidy)
}
\arguments{
\item{scores_temp}{data.frame of fitpoly scores output}

\item{ploidy}{integer defining ploidy}
}
\description{
Interpolation plot
}
