% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plots_overall}
\alias{plots_overall}
\title{Generate plots for overall analysis}
\usage{
plots_overall(est.ploidy.chr_df, filter_diff = 0, filter_corr = 0)
}
\arguments{
\item{est.ploidy.chr_df}{list of data.frames with ploidy estimation information}

\item{filter_diff}{filter by difference on area proportion between first and second place}

\item{filter_corr}{filter by correlation of observed and expected peak position}
}
\description{
Generate plots for overall analysis
}
