% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_main.R
\name{print.hmm_CN}
\alias{print.hmm_CN}
\title{Print method for hmm_CN objects}
\usage{
\method{print}{hmm_CN}(x, ...)
}
\arguments{
\item{x}{An object of class 'hmm_CN'.}

\item{...}{Additional arguments (ignored).}
}
\description{
Prints a summary of the HMM copy-number estimation results, including key parameters.
If the object contains a params_samples list (multi-sample), prints the number of samples.
Otherwise, prints details from the params list.
}
