% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy_est_functions.R
\name{print.qploidy_area_ploidy_estimation}
\alias{print.qploidy_area_ploidy_estimation}
\title{print qploidy_area_ploidy_estimation object}
\usage{
\method{print}{qploidy_area_ploidy_estimation}(x, ...)
}
\arguments{
\item{x}{qploidy_area_ploidy_estimation object}

\item{...}{print parameters}
}
\description{
print qploidy_area_ploidy_estimation object
}
