% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{print.qploidy_interpolation}
\alias{print.qploidy_interpolation}
\title{Print method for object of class 'qploidy_interpolation'}
\usage{
\method{print}{qploidy_interpolation}(x, ...)
}
\arguments{
\item{x}{object of class 'qploidy_interpolation'}

\item{...}{}
}
\value{
printed information about Qploidy interpolation process
}
\description{
Print method for object of class 'qploidy_interpolation'
}
