% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{read_axiom}
\alias{read_axiom}
\title{Converts axiom array summary file to Qploidy and fitpoly input data}
\usage{
read_axiom(
  summary_file,
  ind_names = NULL,
  sd.normalization = FALSE,
  atan = FALSE
)
}
\arguments{
\item{summary_file}{Axiom summary file}

\item{ind_names}{if the summary file columns does not have the desirable sample names, provide a file with two columns: Plate_name: with sample IDs contained in the summary file; Sample_Name: with desirable sample names to be replaced}

\item{sd.normalization}{logical. If TRUE performs standard normalization}

\item{atan}{logical. If TRUE calculates the theta using atan2}
}
\description{
Converts axiom array summary file to Qploidy and fitpoly input data
}
