% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{read_illumina_array}
\alias{read_illumina_array}
\title{read illumina array files. Add suffix _[file number] if more than one file.}
\usage{
read_illumina_array(...)
}
\arguments{
\item{...}{illumina array filename/s}
}
\description{
read illumina array files. Add suffix _[file number] if more than one file.
}
