% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{read_polyOrigin}
\alias{read_polyOrigin}
\title{Extracts genotype probabilities information from polyOrigin polyancentry output file
Code adapted from diaQTL}
\usage{
read_polyOrigin(filename, outstem = NULL)
}
\arguments{
\item{filename}{character with path to polyancentry CSV file}

\item{outstem}{output pedigree filename}
}
\description{
Extracts genotype probabilities information from polyOrigin polyancentry output file
Code adapted from diaQTL
}
