% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{rm_outlier}
\alias{rm_outlier}
\title{Identify outliers using Bonferroni-Holm tests for the adjusted p-values and remove them from the input vector}
\usage{
rm_outlier(plot_data_split_one, alpha = 0.05)
}
\arguments{
\item{plot_data_split_one}{todo}

\item{alpha}{todo}
}
\description{
Identify outliers using Bonferroni-Holm tests for the adjusted p-values and remove them from the input vector
}
