% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inputs.R
\name{simulate_vcf}
\alias{simulate_vcf}
\title{Simulate a VCF file with GT, DP, and AD format fields for 2 chromosomes and 50 samples}
\usage{
simulate_vcf(file_path, seed)
}
\arguments{
\item{file_path}{The path where the simulated VCF file will be saved.}

\item{seed}{The seed for random number generation to ensure reproducibility.}
}
\value{
None. The function writes the simulated VCF content to the specified file.
}
\description{
Simulate a VCF file with GT, DP, and AD format fields for 2 chromosomes and 50 samples
}
