% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{summary_to_fitpoly}
\alias{summary_to_fitpoly}
\title{Convert Summary file to fitpoly format}
\usage{
summary_to_fitpoly(R_all, theta_all)
}
\arguments{
\item{R_all}{data.frame with R values}

\item{theta_all}{data.frame with theta values}

\item{ind.names}{data.frame first column with the plate name and second with sample name}
}
\description{
Convert Summary file to fitpoly format
}
