% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_ttest.R
\name{window_level_test_plot_qploidy}
\alias{window_level_test_plot_qploidy}
\title{Window-level  Depth Comparison Plot}
\usage{
window_level_test_plot_qploidy(
  input,
  selected_samples,
  geno.pos = NULL,
  col2test = c("R", "z"),
  window_size = 1e+06
)
}
\arguments{
\item{input}{Accepts one of:
1. A data.frame with columns: MarkerName, SampleName, Chr, Position, R. Columns Chr and Position are optional, but if they are absent geno.pos object is required. Column with z is optional but required if col2test = "z".
2. An object of class 'qploidy_standardization' (from Qploidy standardization functions)
3. A character string: path to a .tsv[.gz] file with standardized Qploidy data
The function will validate and process the input according to its type.}

\item{selected_samples}{Character vector of sample names to compare against the population.}

\item{geno.pos}{Optional. A data.frame with columns: MarkerName, Chromosome, Position. Required if input data.frame does not contain Chr and Position columns.}

\item{col2test}{Column name to use for depth comparison. Options are "R" or "z".}

\item{window_size}{Size of genomic windows in base pairs (default 1e6).}
}
\value{
A ggplot2 object representing window-level SNP depth comparison.
}
\description{
Generates a scatter plot comparing normalized SNP depth of selected samples against the population mean in genomic windows.
Significance is calculated per window using unpaired t-tests and indicated in the legend.
}
\details{
The function reads standardized SNP depth data from the input, normalizes it per sample,
and aggregates values within windows across chromosomes. It performs unpaired t-tests per window
to compare the selected sample against the population mean. Significance is categorized by direction
(higher or lower than population) and adjusted p-value (FDR).

The resulting ggplot2 scatter plot shows normalized SNP depth per window along the x-axis (megabase),
colored by significance, with facets for each sample and chromosome.
}
\examples{
\dontrun{
window_level_test_plot_qploidy(
  qploidy_object = "standardization.tsv.gz",
  selected_samples = "J432",
  col2test = "z",
  window_size = 1e6
)
}

}
\author{
Josue Chinchilla-Vargas
}
