% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_main.R
\name{write_hmm_CN}
\alias{write_hmm_CN}
\title{Write hmm_CN object to three CSV files}
\usage{
write_hmm_CN(hmm_CN, prefix)
}
\arguments{
\item{hmm_CN}{An object of class 'hmm_CN'.}

\item{prefix}{File prefix for output files (character scalar).}
}
\description{
Writes the three main components of an hmm_CN object (by_window, by_marker, params) to CSV files.
The user must provide a file prefix; files will be named <prefix>_by_window.csv, <prefix>_by_marker.csv, <prefix>_params.csv.
}
