% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{write_qploidy_standardization}
\alias{write_qploidy_standardization}
\title{Write Qploidy Standardization Object to File}
\usage{
write_qploidy_standardization(qploidy_standardization_object, out_filename)
}
\arguments{
\item{qploidy_standardization_object}{An object of class `qploidy_standardization` to be written to file.}

\item{out_filename}{A string specifying the path to the output file where the data will be saved. Specify in the file name the desired extension (e.g., `my_output.csv`).}
}
\value{
None. The function writes the data to the specified file.
}
\description{
This function writes a `qploidy_standardization` object to a specified file.
The output file includes metadata, filtering information, and the standardized dataset.
}
