% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_functions.R
\name{all_resolutions_plots}
\alias{all_resolutions_plots}
\title{Plot graphics for ploidy visual inspection for each resolution
It was made for parallelization purpose}
\usage{
all_resolutions_plots(
  data_standardized,
  sample,
  ploidy,
  centromeres,
  file_name,
  chr
)
}
\arguments{
\item{data_standardized}{result from standardization function}

\item{sample}{sample name}

\item{ploidy}{sample ploidy if known}

\item{centromeres}{vector with centromeres positions}

\item{file_name}{character defining the output file path and name}

\item{chr}{vector containing which chromosomes should be included}
}
\description{
Plot graphics for ploidy visual inspection for each resolution
It was made for parallelization purpose
}
