% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{clean_summary}
\alias{clean_summary}
\title{Edit Axiom Summary file}
\usage{
clean_summary(summary_df)
}
\arguments{
\item{summary_df}{data.frame with A and B probes intensities}
}
\value{
list with A and B probes
}
\description{
Remove consecutive A allele probe from the summary file
}
