% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_counts_functions.R
\name{count_breaks_df}
\alias{count_breaks_df}
\title{Count number of recombination breakpoints in parents gametes}
\usage{
count_breaks_df(homoprob, aneuploids = NULL, P1 = "P1", P2 = "P2", inds = NULL)
}
\arguments{
\item{P1}{string defining parent 1 name}

\item{P2}{string defining parent 2 name}

\item{inds}{vector with individuals to be evaluated}

\item{object}{of class mappoly.homoprob}
}
\description{
Count number of recombination breakpoints in parents gametes
}
