% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dosages.R
\name{export_VCF}
\alias{export_VCF}
\title{Export hmm_dosage_calls to VCF}
\usage{
export_VCF(hmm_dosage_calls, file)
}
\arguments{
\item{hmm_dosage_calls}{A data.frame with columns: MarkerName, SampleName, Chr, Position, X, Y, CN_call, post_max_CN, dosage, post_max_dosage, BAF, z}

\item{file}{Path to output VCF file}
}
\value{
The path to the written VCF file (invisible)
}
\description{
Converts a hmm_dosage_calls object (data.frame) to a VCF file with GT, CN, AD, BAF, Z, post_max_CN, and post_max_dosage in FORMAT.
}
\details{
The VCF INFO field will contain the mode of CN_call values for each marker. The FORMAT field will include GT (genotype), CN (copy number call), AD (allelic depths), BAF, Z, PMC (posterior max CN), and PMD (posterior max dosage). Genotypes are assigned based on the dosage and CN_call, with 1 representing the alternate allele and 0 representing the reference allele.
}
