% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy_est.R
\name{get_aneuploids}
\alias{get_aneuploids}
\title{indexes for aneuploids}
\usage{
get_aneuploids(ploidy_df)
}
\arguments{
\item{ploidy_df}{ploidy table (chromosome in columns and individuals in rows)}
}
\value{
A logical vector where each element corresponds to an individual in the
        input ploidy table. The value is `TRUE` if the individual is identified
        as potentially aneuploid, and `FALSE` otherwise.
}
\description{
indexes for aneuploids
}
