% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{get_centers}
\alias{get_centers}
\title{Define cluster centers}
\usage{
get_centers(
  ratio_geno,
  ploidy,
  n.clusters.thr = NULL,
  type = c("intensities", "counts"),
  rm_outlier = TRUE
)
}
\arguments{
\item{ratio_geno}{ratio and genotype information for each marker and sample. data.frame with columns: 1) mks; 2) ind; 3) theta; 4) geno.}

\item{ploidy}{integer defining ploidy}

\item{n.clusters.thr}{minimum number of clusters required. If clusters < ploidy + 1, the missing clusters will be imputed}

\item{type}{if array alleles intensities data "intensities", if read counts data "counts"}

\item{rm_outlier}{if TRUE, it remove the outlier markers before defining the cluster centers}
}
\description{
Define cluster centers
}
