% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{get_R_theta}
\alias{get_R_theta}
\title{Get R and theta values from summary file with option to do standard normalization by plate and markers}
\usage{
get_R_theta(cleaned_summary, ind.names, sd.normalization = FALSE, atan = FALSE)
}
\arguments{
\item{cleaned_summary}{summary object from clean_summary function. List with A and B intensities.}

\item{ind.names}{data.frame with individual names correspondence}

\item{sd.normalization}{logical. If TRUE performs standard normalization}

\item{atan}{logical. If TRUE calculates the theta using atan2}
}
\description{
Get R and theta values from summary file with option to do standard normalization by plate and markers
}
