% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{interpolate_BAFs}
\alias{interpolate_BAFs}
\title{Performs interpolation and returns normalized BAFs}
\usage{
interpolate_BAFs(
  data = NULL,
  genos = NULL,
  geno.pos = NULL,
  threshold.missing.geno = 0.5,
  ploidy.interpolation = NULL,
  threshold.n.clusters = NULL,
  n.cores = 1,
  baf_file_name,
  verbose = TRUE
)
}
\arguments{
\item{data}{data.frame with columns:1) MarkeName: markers IDs;2) SampleName: Samples IDs;3) X: reference allele intensities or counts;4) Y: alternative allele intensities or counts,5) R: sum of the intensities; and 6)ratio: Y/(X+Y)}

\item{genos}{data.frame with genotype information for individuals to be used as reference for interpolation. We suggest to select for these individuals that are euploid and all have same ploidy. For array technologies, we suggest using fitpoly for obtaining the dosages, and for sequencing technologies, updog. This file has as columns: 1) MarkerName: markers IDs; 2) SampleName: Samples IDs; 3) geno: dosage.}

\item{geno.pos}{data.frame with columns: 1) MarkerName: markers IDs; Chromosome: chromosome where the marker is located; Position: position on the chromosome the marker is located (bp).}

\item{threshold.missing.geno}{fraction of missing genotype information allowed by marker. Markers with higher fraction are discarted}
}
\description{
Performs interpolation and returns normalized BAFs
}
