% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{par_fitpoly_interpolation}
\alias{par_fitpoly_interpolation}
\title{Clusterization using fitpoly}
\usage{
par_fitpoly_interpolation(scores_temp, ploidy, n.clusters.thr = 0)
}
\arguments{
\item{scores_temp}{fitpoly scores output file}

\item{ploidy}{integer defining ploidy}

\item{n.clusters.thr}{minimum number of clusters required. If clusters < ploidy + 1, the missing clusters will be imputed.}

\item{plot}{logical to create or not the interpolation plot}
}
\description{
Clusterization using fitpoly
}
