% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{par_fitpoly_interpolation}
\alias{par_fitpoly_interpolation}
\title{Clusterization using fitpoly}
\usage{
par_fitpoly_interpolation(scores_temp, ploidy, plot = TRUE)
}
\description{
Clusterization using fitpoly
}
