% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pascalTriangle}
\alias{pascalTriangle}
\title{Pascal triangle to define the expected peaks for each ploidy}
\usage{
pascalTriangle(h)
}
\arguments{
\item{h}{ploidy value}
}
\description{
Pascal triangle to define the expected peaks for each ploidy
}
