% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy_est_functions.R
\name{plot_baf_hist}
\alias{plot_baf_hist}
\title{Plot BAF histogram}
\usage{
plot_baf_hist(data_sample, area_single, ploidy, colors, add_lines)
}
\arguments{
\item{data_sample}{data.frame with BAF and genomic position information}

\item{area_single}{area around the expected peak to be considered}

\item{ploidy}{expected ploidy}

\item{colors}{add area colors}

\item{add_lines}{add expected peaks lines}
}
\description{
Plot BAF histogram
}
