% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_functions.R
\name{plot_baf_hist}
\alias{plot_baf_hist}
\title{Plot BAF histogram}
\usage{
plot_baf_hist(
  data_sample,
  area_single,
  ploidy,
  colors = FALSE,
  add_estimated_peaks = TRUE,
  add_expected_peaks = FALSE,
  BAF_hist_overall = FALSE,
  ratio = FALSE,
  rm_homozygous = FALSE,
  font_size = 12
)
}
\arguments{
\item{data_sample}{data.frame with BAF and genomic position information}

\item{area_single}{area around the expected peak to be considered}

\item{ploidy}{expected ploidy}

\item{colors}{add area colors}

\item{add_estimated_peaks}{add expected peaks lines}

\item{add_expected_peaks}{add estimated peaks}

\item{BAF_hist_overall}{if TRUE it plots the BAF histogram for entire genome}

\item{ratio}{if TRUE plot the raw ratio}

\item{rm_homozygous}{if TRUE removes the homozygous calls}

\item{font_size}{graphic labels font size}
}
\description{
Plot BAF histogram
}
