% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_functions.R
\name{plot_check}
\alias{plot_check}
\title{3D plot to check raw data allele intensities (if array) or depth (if sequencing) data properties
Qploidy can be used if that is not much variation among the individuals for the same marker.
So we expect that the overall structure of the resulting graphic is a ramp.}
\usage{
plot_check(data_qploidy, R_lim = NULL, n = 3000)
}
\arguments{
\item{data_qploidy}{data.frame input for Qploidy}

\item{R_lim}{limit the sum of intensities/depth axis to adjust the dimentions in case there are outliers}

\item{n}{subset size}
}
\description{
3D plot to check raw data allele intensities (if array) or depth (if sequencing) data properties
Qploidy can be used if that is not much variation among the individuals for the same marker.
So we expect that the overall structure of the resulting graphic is a ramp.
}
