% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_qploidy_standardization}
\alias{plot_qploidy_standardization}
\title{Plot method for object of class 'qploidy_standardization'}
\usage{
plot_qploidy_standardization(
  x,
  sample = NULL,
  chr = NULL,
  type = c("all", "het", "BAF", "zscore", "BAF_hist", "BAF_hist_overall",
    "Ratio_hist_overall"),
  area_single = 0.75,
  ploidy = 4,
  dot.size = 1,
  font_size = 12,
  add_estimated_peaks = FALSE,
  add_expected_peaks = FALSE,
  centromeres = NULL,
  add_centromeres = FALSE,
  colors = FALSE,
  window_size = 2e+06,
  het_interval = 0.1,
  rm_homozygous = FALSE,
  ...
)
}
\arguments{
\item{x}{object of class 'qploidy_standardization'}

\item{sample}{character indicating sample ID}

\item{chr}{chromosome index to be plotted}

\item{type}{character defining the graphic type. "all" plots all graphics,
it is equivalent to "het","ratio","BAF","zscore","BAF_hist", "BAF_hist_overall".
"het" plots the heterozygous locus counts inside genomic positions window defined in 'window_size'.
"ratio" plots raw Y/(X+Y) or alternative count/(alternative counts + reference counts).
"BAF" plots standardized ratios. "zscore" is the smoothed conditional means curve of standardized sum of intensities/counts.
"BAF_hist" is the histogram of standardized ratios. "BAF_hist_overall" add the histogram including all markers.}

\item{area_single}{area around the expected peak to be considered}

\item{ploidy}{expected ploidy}

\item{dot.size}{dot size}

\item{font_size}{graphic labels font size}

\item{add_estimated_peaks}{add expected peaks lines}

\item{add_expected_peaks}{add estimated peaks}

\item{centromeres}{centromeres position for each chromosome}

\item{add_centromeres}{if TRUE add a vertical line at the centromere position}

\item{colors}{add area colors}

\item{window_size}{genomic position window to calculate the number of heterozygous locus}

\item{het_interval}{interval to be considered as heterozygous (heterozygous ratio > 1 - het_interval and heterozygous ratio < 0 + het_interval)}

\item{rm_homozygous}{if TRUE remove the homozygous for the BAF histogram plots}

\item{...}{plot parameters}
}
\value{
printed information about Qploidy standardization process
}
\description{
Plot method for object of class 'qploidy_standardization'
}
