% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{qploidy_read_vcf}
\alias{qploidy_read_vcf}
\title{convert vcf file to Qploidy data}
\usage{
qploidy_read_vcf(vcf_file)
}
\arguments{
\item{vcf}{path to vcf file}
}
\description{
convert vcf file to Qploidy data
}
