% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{read_illumina_array}
\alias{read_illumina_array}
\title{Read Illumina Array Files}
\usage{
read_illumina_array(...)
}
\arguments{
\item{...}{One or more Illumina array filenames.}
}
\value{
A data frame containing the processed Illumina array data.
}
\description{
This function reads Illumina array files and processes them into a format
suitable for Qploidy analysis. It adds a suffix to sample IDs if multiple
files are provided.
}
