% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{read_polyOrigin}
\alias{read_polyOrigin}
\title{Extracts genotype probabilities information from polyOrigin polyancentry output file
Code adapted from diaQTL}
\usage{
read_polyOrigin(filename)
}
\arguments{
\item{filename}{character with path to polyancentry CSV file}
}
\description{
Extracts genotype probabilities information from polyOrigin polyancentry output file
Code adapted from diaQTL
}
