% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{read_qploidy_standardization}
\alias{read_qploidy_standardization}
\title{Read file generated by standardize function}
\usage{
read_qploidy_standardization(qploidy_standardization_file)
}
\arguments{
\item{qploidy_standardization_file}{path to file generated by standardize function}
}
\description{
Read file generated by standardize function
}
