% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation_functions.R
\name{standardize}
\alias{standardize}
\title{Performs standardization and returns BAF and z score}
\usage{
standardize(
  data = NULL,
  genos = NULL,
  geno.pos = NULL,
  threshold.missing.geno = 0.9,
  threshold.geno.prob = 0.8,
  ploidy.standardization = NULL,
  threshold.n.clusters = NULL,
  n.cores = 1,
  out_filename = NULL,
  type = "intensities",
  multidog_obj = NULL,
  parallel.type = "PSOCK",
  verbose = TRUE,
  rm_outlier = TRUE,
  cluster_median = TRUE
)
}
\arguments{
\item{data}{data.frame with columns: 1) MarkeName: markers IDs; 2) SampleName: Samples IDs; 3) X: reference allele intensities or counts; 4) Y: alternative allele intensities or counts; 5) R: sum of the intensities; and 6) ratio: Y/(X+Y)}

\item{genos}{data.frame with genotype information for individuals to be used as reference for standardization. We suggest to select for these individuals that are euploid and all have same ploidy. For array technologies, we suggest using fitpoly for obtaining the dosages, and for sequencing technologies, updog. This file has as columns: 1) MarkerName: markers IDs; 2) SampleName: Samples IDs; 3) geno: dosage.}

\item{geno.pos}{data.frame with columns: 1) MarkerName: markers IDs; 2) Chromosome: chromosome where the marker is located; 3) Position: position on the chromosome the marker is located (bp).}

\item{threshold.missing.geno}{fraction of missing genotype information allowed by marker. Markers with higher fraction are discarted}

\item{threshold.geno.prob}{minimum genotype probability allowed. Genotypes with lower probability will be replaced by NA}

\item{ploidy.standardization}{ploidy of the reference samples defined in `genos`}

\item{threshold.n.clusters}{minimum number of dosage clusters (heterozygous classes) to account with the marker for standardization}

\item{n.cores}{number of cores to be used in parallelized processes}

\item{out_filename}{output file name}

\item{type}{method to determine the clusters centers for each marker. It can be "intensities" if array data, "counts" if sequencing data or "updog" if multidog object is provided in multidog_obj argument}

\item{multidog_obj}{object of class multidog from updog package analysis}

\item{parallel.type}{FORK or PSOCK}

\item{verbose}{If TRUE display informative messages}

\item{rm_outlier}{if TRUE remove outliers while standardizing}

\item{cluster_median}{if TRUE use median to define the cluster centroids}
}
\description{
Performs standardization and returns BAF and z score
}
