% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{summary_to_fitpoly}
\alias{summary_to_fitpoly}
\title{Convert Summary file to fitpoly format}
\usage{
summary_to_fitpoly(cleaned_summary, ind.names, geno.pos)
}
\arguments{
\item{cleaned_summary}{list with A and B intensities}

\item{ind.names}{data.frame first column with the plate name and second with sample name}
}
\description{
Convert Summary file to fitpoly format
}
